requirements-lite.txt
# 基础依赖
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.8.0
seaborn>=0.13.0

# 音视频处理
opencv-python>=4.8.0
# 音频处理 (简化版，移除了PyAudio)
sounddevice>=0.4.6
librosa>=0.10.1

# 机器学习与深度学习 (简化版，移除了face-recognition和deepface)
scikit-learn>=1.3.2
mediapipe>=0.10.9

# 用户输入监控
pynput>=1.7.6
pygetwindow>=0.0.9
pillow>=10.1.0

# Web界面
flask>=2.3.3
dash>=2.13.0
dash-bootstrap-components>=1.5.0
plotly>=5.18.0
flask-socketio>=5.3.6
python-socketio>=5.10.0

# 工具类
python-dotenv>=1.0.0
tqdm>=4.65.0
